refactor(maestro): reworked head naming
