+++ packages/maestro/src/storybook/head.ts
@@ -1,4 +1,4 @@
-import { HEAD } from "./templates/head";
+import { HEAD } from "./templates/drupal-global";
 
 function head(previous: string) {
   return `
